more refactoring, ready for new functions
