+++ backend-services/database/core/database.py
@@ -11,11 +11,8 @@ import requests
 import ujson as json
 from artemis_utils import get_hash
 from artemis_utils import get_logger
-from artemis_utils import ping_redis
-from artemis_utils import purge_redis_eph_pers_keys
 from artemis_utils import RABBITMQ_URI
 from artemis_utils import REDIS_HOST
-from artemis_utils import redis_key
 from artemis_utils import REDIS_PORT
 from artemis_utils import WITHDRAWN_HIJACK_THRESHOLD
 from artemis_utils.db_util import DB
@@ -28,6 +25,9 @@ from artemis_utils.envvars import DB_USER
 from artemis_utils.envvars import HISTORIC
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
+from artemis_utils.redis import ping_redis
+from artemis_utils.redis import purge_redis_eph_pers_keys
+from artemis_utils.redis import redis_key
 from kombu import Connection
 from kombu import Producer
 from kombu import uuid
