+++ backend-services/notifier/core/notifier.py
@@ -8,8 +8,8 @@ from typing import NoReturn
 import requests
 import ujson as json
 from artemis_utils import get_logger
-from artemis_utils import hijack_log_field_formatter
 from artemis_utils import RABBITMQ_URI
+from artemis_utils.logging import hijack_log_field_formatter
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
 from kombu import Connection
