+++ docs/loggingconf.md
@@ -17,9 +17,9 @@ cd /var/log/artemis
 Both containers use [Python's logging library](https://docs.python.org/3/library/logging.html) and can be configured as such. The tool supports SMTP, SMPTS and SYSLOG handlers for the loggers, which can be defined as follows:
 ```
 smtps_handler:
-    # artemis_utils.TLSSMTPHandler for TLS
-    # artemis_utils.SSLSMTPHandler for SSL
-    class: artemis_utils.SSLSMTPHandler
+    # artemis_utils.logging.TLSSMTPHandler for TLS
+    # artemis_utils.logging.SSLSMTPHandler for SSL
+    class: artemis_utils.logging.SSLSMTPHandler
     level: INFO
     formatter: simple
     mailhost:
