+++ monitor-services/exabgptap/core/exabgp_client.py
@@ -9,15 +9,15 @@ import requests
 import ujson as json
 from artemis_utils import get_ip_version
 from artemis_utils import get_logger
-from artemis_utils import key_generator
-from artemis_utils import normalize_msg_path
-from artemis_utils import ping_redis
 from artemis_utils import RABBITMQ_URI
 from artemis_utils import REDIS_HOST
 from artemis_utils import REDIS_PORT
 from artemis_utils.envvars import MON_TIMEOUT_LAST_BGP_UPDATE
 from artemis_utils.rabbitmq import create_exchange
+from artemis_utils.redis import ping_redis
+from artemis_utils.updates import key_generator
 from artemis_utils.updates import MformatValidator
+from artemis_utils.updates import normalize_msg_path
 from kombu import Connection
 from kombu import Producer
 from socketIO_client import BaseNamespace
