+++ other/monitor_deprecated.py
@@ -9,7 +9,6 @@ from artemis_utils import dump_json
 from artemis_utils import exception_handler
 from artemis_utils import get_ip_version
 from artemis_utils import get_logger
-from artemis_utils import ping_redis
 from artemis_utils import RABBITMQ_URI
 from artemis_utils import REDIS_HOST
 from artemis_utils import REDIS_PORT
@@ -17,6 +16,7 @@ from artemis_utils import search_worst_prefix
 from artemis_utils.envvars import MON_TIMEOUT_LAST_BGP_UPDATE
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
+from artemis_utils.redis import ping_redis
 from artemis_utils.translations import translate_rfc2622
 from kombu import Connection
 from kombu import Consumer
