+++ testing/autoconf/configs/logging.yaml
@@ -31,9 +31,9 @@ handlers:
 
     # for email
     # smtps_handler:
-    #     # artemis_utils.TLSSMTPHandler for TLS
-    #     # artemis_utils.SSLSMTPHandler for SSL
-    #     class: artemis_utils.TLSSMTPHandler
+    #     # artemis_utils.logging.TLSSMTPHandler for TLS
+    #     # artemis_utils.logging.SSLSMTPHandler for SSL
+    #     class: artemis_utils.logging.TLSSMTPHandler
     #     level: INFO
     #     formatter: simple
     #     mailhost:
