+++ utils/artemis_utils/updates.py
@@ -1,4 +1,5 @@
 # functions for update processing
+import copy
 from datetime import datetime
 from datetime import timedelta
 from ipaddress import ip_network as str2ip
@@ -6,6 +7,20 @@ from typing import List
 from typing import Tuple
 
 from envvars import HISTORIC
+
+from . import get_hash
+
+
+def key_generator(msg):
+    msg["key"] = get_hash(
+        [
+            msg["prefix"],
+            msg["path"],
+            msg["type"],
+            "{0:.6f}".format(msg["timestamp"]),
+            msg["peer_asn"],
+        ]
+    )
 
 
 class MformatValidator:
@@ -152,3 +167,68 @@ def clean_as_path(path: List[int]) -> List[int]:
     if is_loopy:
         clean_path = __clean_loops(clean_path)
     return clean_path
+
+
+def __decompose_path(path):
+
+    # first do an ultra-fast check if the path is a normal one
+    # (simple sequence of ASNs)
+    str_path = " ".join(map(str, path))
+    if "{" not in str_path and "[" not in str_path and "(" not in str_path:
+        return [path]
+
+    # otherwise, check how to decompose
+    decomposed_paths = []
+    for hop in path:
+        hop = str(hop)
+        # AS-sets
+        if "{" in hop:
+            decomposed_hops = hop.lstrip("{").rstrip("}").split(",")
+        # AS Confederation Set
+        elif "[" in hop:
+            decomposed_hops = hop.lstrip("[").rstrip("]").split(",")
+        # AS Sequence Set
+        elif "(" in hop or ")" in hop:
+            decomposed_hops = hop.lstrip("(").rstrip(")").split(",")
+        # simple ASN
+        else:
+            decomposed_hops = [hop]
+        new_paths = []
+        if not decomposed_paths:
+            for dec_hop in decomposed_hops:
+                new_paths.append([dec_hop])
+        else:
+            for prev_path in decomposed_paths:
+                if "(" in hop or ")" in hop:
+                    new_path = prev_path + decomposed_hops
+                    new_paths.append(new_path)
+                else:
+                    for dec_hop in decomposed_hops:
+                        new_path = prev_path + [dec_hop]
+                        new_paths.append(new_path)
+        decomposed_paths = new_paths
+    return decomposed_paths
+
+
+def normalize_msg_path(msg):
+    msgs = []
+    path = msg["path"]
+    msg["orig_path"] = None
+    if isinstance(path, list):
+        dec_paths = __decompose_path(path)
+        if not dec_paths:
+            msg["path"] = []
+            msgs = [msg]
+        elif len(dec_paths) == 1:
+            msg["path"] = list(map(int, dec_paths[0]))
+            msgs = [msg]
+        else:
+            for dec_path in dec_paths:
+                copied_msg = copy.deepcopy(msg)
+                copied_msg["path"] = list(map(int, dec_path))
+                copied_msg["orig_path"] = path
+                msgs.append(copied_msg)
+    else:
+        msgs = [msg]
+
+    return msgs
